Aula 2

+++ Aula 2/src/Programa4.java
@@ -3,11 +3,11 @@ import java.util.Scanner;
 public class Programa4 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
-        int num1 = input.nextInt();
-        int num2 = input.nextInt();
-        int num3 = input.nextInt();
+        float num1 = input.nextFloat();
+        float num2 = input.nextFloat();
+        float num3 = input.nextFloat();
 
-        int media = (num1 + num2 + num3) / 3;
+        float media = (num1 + num2 + num3) / 3;
 
         if (media >= 7) {
             System.out.println("Sua média é maior ou igual que sete!\nMédia: " + media);
